requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# Environment Variables
python-decouple==3.8

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# CORS Headers
django-cors-headers==4.3.1

# Filtering
django-filter==23.5

# Utilities
Pillow==10.1.0

# Development Tools (optional but recommended)
ipython==8.18.1
django-extensions==3.2.3

